Create skutofile
This was written more to debug what my program is pulling but works well enough as a stand alone script.  Basically this program will connect to your store and parse the xml to a dictionary with the sku as the key and the qty level as the value 1. Prints dictionary length after each page which I find helpful but can be skipped by commenting out lines 28 & 29.

Once it has pulled all the products it writes it to a csv, which has to columns the first is the sku the second qty.