Update bcorders.py
This program pulls all of the order id's since it was last ran.  It confirms that orders have a captured payment because incomplete orders are listed in the xml. Then the script pulls the sku's and quantity of the order into a dictionary and if the same item is sold multiple times it adds the new quantity value to the existing sku dictionary entry